@@ -1,37 +1,41 @@
-from pydantic import BaseModel, Field, field_validator, ConfigDict
-from typing import List, Dict, Any, Optional
-import re
+from typing import Any
+
+from pydantic import BaseModel, Field, field_validator
+
 
 class PredictionRequest(BaseModel):
-    model_config = ConfigDict(
-        json_schema_extra={
+    """Request model for batch prediction"""
+
+    model_config = {
+        "json_schema_extra": {
             "example": {
                 "models": ["fraud_detection:v1", "credit_score:v1"],
                 "entities": {"cust_no": ["X123456", "1002"]},
-                "event_timestamp": 1751429485000
+                "event_timestamp": 1751429485000,
             }
         }
+    }
+
+    models: list[str] = Field(
+        ..., description="List of model names with versions in format 'model_name:version'"
     )
-    
-    models: List[str] = Field(..., description="List of model names with versions in format 'model_name:version'")
-    entities: Dict[str, List[Any]] = Field(..., description="Entity mapping with IDs")
-    event_timestamp: Optional[int] = Field(None, description="Event timestamp in milliseconds GMT+0")
-    
-    @field_validator('models')
+    entities: dict[str, list[Any]] = Field(..., description="Entity mapping with IDs")
+    event_timestamp: int | None = Field(None, description="Event timestamp in milliseconds GMT+0")
+
+    @field_validator("models")
     @classmethod
     def validate_models(cls, v):
         if not isinstance(v, list):
-            raise ValueError("models must be a list")
-        
+            raise TypeError("models must be a list")
+
         # Allow empty list and any model format - validation happens at service level
-        # This allows the API to accept requests and handle errors gracefully
         return v
-    
-    @field_validator('entities')
+
+    @field_validator("entities")
     @classmethod
     def validate_entities(cls, v):
         if not isinstance(v, dict):
-            raise ValueError("entities must be a dictionary")
+            raise TypeError("entities must be a dictionary")
         if not v:
-            raise ValueError("entities cannot be empty")
+            raise ValueError("entities cannot be empty")  # noqa: EM101
         return v
